+++ cogeo_mosaic/backends/sqlite.py
@@ -273,7 +273,11 @@ class SQLiteBackend(BaseBackend):
         """Find assets."""
         mercator_tile = mercantile.Tile(x=x, y=y, z=z)
         quadkeys = find_quadkeys(mercator_tile, self.quadkey_zoom)
-        return list(itertools.chain.from_iterable([self._fetch(qk) for qk in quadkeys]))
+        return list(
+            dict.fromkeys(
+                itertools.chain.from_iterable([self._fetch(qk) for qk in quadkeys])
+            )
+        )
 
     @property
     def _quadkeys(self) -> List[str]: