+++ tests/test_backends_utils.py
@@ -5,7 +5,6 @@ import os
 import re
 
 import mercantile
-import pytest
 
 from cogeo_mosaic.backends import utils
 
@@ -63,21 +62,3 @@ def test_find_quadkeys():
         "0302303303",
         "0302303302",
     ]
-
-
-def test_get_assets_from_json():
-    """Get assets list."""
-    qkz = mosaic_content.get("quadkey_zoom") or mosaic_content.get("minzoom")
-    with pytest.warns(DeprecationWarning):
-        assert (
-            len(utils.get_assets_from_json(mosaic_content["tiles"], qkz, 150, 182, 9))
-            == 2
-        )
-        assert (
-            len(utils.get_assets_from_json(mosaic_content["tiles"], qkz, 147, 182, 9))
-            == 1
-        )
-        assert (
-            len(utils.get_assets_from_json(mosaic_content["tiles"], qkz, 147, 182, 12))
-            == 0
-        )