+++ cogeo_mosaic/__init__.py
@@ -1,3 +1,3 @@
 """Cogeo_mosaic."""
 
-__version__ = "3.0.2"
+__version__ = "4.0.0a0"