+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Release Notes
 
-## 4.0.0a0 (TBD)
+## 4.0.0a0 (2021-10-20)
 
 * update morecantile requirement to >= 3.0
 * update rio-tiler requirement to >= 3.0 and update Backend's properties